fix(config): move into one file
